Validation class added. Need tests
